Move WireMock to dedicated file in UITestsFoundation
