+++ WordPress/UITestsFoundation/BaseScreen.swift
@@ -112,22 +112,6 @@ extension BaseScreen {
     }
 }
 
-public class WireMock {
-    private static let hostInfoPlistKey = "WIREMOCK_HOST"
-    private static let portInfoPlistKey = "WIREMOCK_PORT"
-
-    public static func URL() -> Foundation.URL {
-        let host = infoPlistEntry(key: hostInfoPlistKey)
-        let port = infoPlistEntry(key: portInfoPlistKey)
-        return Foundation.URL(string: "http://\(host):\(port)/")!
-    }
-
-    private static func infoPlistEntry(key: String) -> String {
-        let plistUrl = Bundle(for: WireMock.self).url(forResource: "Info", withExtension: "plist")!
-        return NSDictionary(contentsOf: plistUrl)![key] as! String
-    }
-}
-
 // TODO: This might be better suited in the XCUITestHelpers frameworks.
 public extension XCUIElementQuery {
